Some external commands refactor [skip ci]
