+++ src/network/protocols/command_manager.hpp
@@ -40,6 +40,11 @@
 
 #include "network/protocols/command_voting.hpp"
 #include "network/protocols/command_permissions.hpp"
+#include "utils/command_manager/auth_resource.hpp"
+#include "utils/command_manager/command.hpp"
+#include "utils/command_manager/context.hpp"
+#include "utils/command_manager/file_resource.hpp"
+#include "utils/command_manager/text_resource.hpp"
 #include "utils/enum_extended_reader.hpp"
 #include "utils/lobby_context.hpp"
 #include "utils/set_typo_fixer.hpp"
@@ -55,47 +60,7 @@ class STKPeer;
 
 class CommandManager: public LobbyContextComponent
 {
-public:
-    enum ModeScope: int
-    {
-        MS_DEFAULT = 1,
-        MS_SOCCER_TOURNAMENT = 2
-        // add more powers of two if needed
-    };
-
-    enum StateScope: int
-    {
-        SS_LOBBY = 1,
-        SS_INGAME = 2,
-        SS_ALWAYS = SS_LOBBY | SS_INGAME
-    };
-
 private:
-    struct FileResource
-    {
-        std::string m_file_name;
-        uint64_t m_interval;
-        uint64_t m_last_invoked;
-        std::string m_contents;
-
-        FileResource(std::string file_name = "", uint64_t interval = 0);
-
-        void read();
-
-        std::string get();
-    };
-
-    struct AuthResource
-    {
-        std::string m_secret;
-        std::string m_server;
-        std::string m_link_format;
-
-        AuthResource(std::string secret = "", std::string server = "",
-            std::string link_format = "");
-
-        std::string get(const std::string& username, int online_id) const;
-    };
 
     static EnumExtendedReader permission_reader;
     static EnumExtendedReader mode_scope_reader;
@@ -105,115 +70,6 @@ private:
 
     template<typename T>
     void add_to_queue(int x, int mask, bool to_front, std::string& s) const;
-
-    struct Command;
-
-    struct Context
-    {
-        ServerLobby* m_lobby;
-
-        Event* m_event;
-
-        std::weak_ptr<STKPeer> m_peer;
-
-        std::weak_ptr<STKPeer> m_target_peer;
-
-        std::vector<std::string> m_argv;
-
-        std::string m_cmd;
-
-        std::weak_ptr<Command> m_command;
-
-        int m_user_permissions;
-
-        int m_acting_user_permissions;
-
-        bool m_voting;
-
-        Context(ServerLobby* lobby, Event* event, std::shared_ptr<STKPeer> peer):
-                m_lobby(lobby),
-                m_event(event), m_peer(peer), m_target_peer(peer), m_argv(),
-                m_cmd(""), m_user_permissions(0), m_acting_user_permissions(0), m_voting(false) {}
-
-        Context(ServerLobby* lobby, Event* event, std::shared_ptr<STKPeer> peer,
-            std::vector<std::string>& argv, std::string& cmd,
-            int user_permissions, int acting_user_permissions, bool voting):
-                m_lobby(lobby),
-                m_event(event), m_peer(peer), m_target_peer(peer), m_argv(argv),
-                m_cmd(cmd), m_user_permissions(user_permissions),
-                m_acting_user_permissions(acting_user_permissions), m_voting(voting) {}
-
-        std::shared_ptr<STKPeer> peer();
-        std::shared_ptr<STKPeer> peerMaybeNull();
-        std::shared_ptr<STKPeer> actingPeer();
-        std::shared_ptr<STKPeer> actingPeerMaybeNull();
-        std::shared_ptr<Command> command();
-
-        void say(const std::string& s);
-    };
-
-    struct CommandDescription
-    {
-        std::string m_usage;
-        std::string m_permissions;
-        std::string m_description;
-        CommandDescription(std::string usage = "", std::string permissions = "",
-            std::string description = ""): m_usage(usage),
-            m_permissions(permissions), m_description(description) {}
-
-        std::string getUsage() const { return "Usage: " + m_usage; }
-
-        std::string getHelp() const
-        {
-            return "Usage: " + m_usage
-                + "\nAvailable to: " + m_permissions
-                + "\n" + m_description;
-        }
-    };
-
-    struct Command
-    {
-        std::string m_name;
-
-        std::string m_prefix_name;
-
-        void (CommandManager::*m_action)(Context& context);
-
-        int m_permissions;
-
-        int m_mode_scope;
-
-        int m_state_scope;
-
-        bool m_omit_name;
-
-        CommandDescription m_description;
-
-        std::weak_ptr<Command> m_parent;
-
-        std::map<std::string, std::weak_ptr<Command>> m_name_to_subcommand;
-
-        std::vector<std::shared_ptr<Command>> m_subcommands;
-
-        SetTypoFixer m_stf_subcommand_names;
-
-        Command() {}
-
-        Command(std::string name,
-                void (CommandManager::*f)(Context& context),
-                int permissions = UP_EVERYONE,
-                int mode_scope = MS_DEFAULT, int state_scope = SS_ALWAYS);
-
-        void changeFunction(void (CommandManager::*f)(Context& context))
-                                                              { m_action = f; }
-        void changePermissions(int permissions = UP_EVERYONE,
-                                int mode_scope = MS_DEFAULT,
-                                int state_scope = SS_ALWAYS);
-
-        std::string getUsage() const       { return m_description.getUsage(); }
-        std::string getHelp() const         { return m_description.getHelp(); }
-        std::string getFullName() const               { return m_prefix_name; }
-    };
 
 private:
 
@@ -225,7 +81,7 @@ private:
 
     std::multiset<std::string> m_users;
 
-    std::map<std::string, std::string> m_text_response;
+    std::map<std::string, TextResource> m_text_response;
 
     std::map<std::string, FileResource> m_file_resources;
 
@@ -384,15 +240,21 @@ public:
 
     void handleCommand(Event* event, std::shared_ptr<STKPeer> peer);
 
-    template<typename T>
-    void addTextResponse(std::string key, T&& value)
-                                              { m_text_response[key] = value; }
+    // template<typename T>
+    // void addTextResponse(std::string key, T&& value)
+    //                                           { m_text_response[key] = value; }
 
-    void addFileResource(std::string key, std::string file, uint64_t interval)
-                      { m_file_resources[key] = FileResource(file, interval); }
+    void addTextResponse(std::string key, const TextResource& resource)
+                                           { m_text_response[key] = resource; }
 
-    void addAuthResource(std::string key, std::string secret, std::string server, std::string link_format)
-         { m_auth_resources[key] = AuthResource(secret, server, link_format); }
+    void addTextResponse(std::string key, TextResource&& resource)
+                                { m_text_response[key] = std::move(resource); }
+
+    void addFileResource(std::string key, const FileResource& resource)
+                                          { m_file_resources[key] = resource; }
+
+    void addAuthResource(std::string key, const AuthResource& resource)
+                                          { m_auth_resources[key] = resource; }
 
     void addUser(std::string& s);
