+++ src/utils/command_manager/auth_resource.hpp
@@ -1,0 +1,38 @@
+//
+//  SuperTuxKart - a fun racing game with go-kart
+//  Copyright (C) 2025 kimden
+//
+//  This program is free software; you can redistribute it and/or
+//  modify it under the terms of the GNU General Public License
+//  as published by the Free Software Foundation; either version 3
+//  of the License, or (at your option) any later version.
+//
+//  This program is distributed in the hope that it will be useful,
+//  but WITHOUT ANY WARRANTY; without even the implied warranty of
+//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//  GNU General Public License for more details.
+//
+//  You should have received a copy of the GNU General Public License
+//  along with this program; if not, write to the Free Software
+//  Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
+
+#ifndef AUTH_RESOURCE_HPP
+#define AUTH_RESOURCE_HPP
+
+#include <string>
+#include "utils/command_manager/abstract_resource.hpp"
+
+class AuthResource: public AbstractResource
+{
+private:
+    std::string m_secret;
+    std::string m_server;
+    std::string m_link_format;
+
+public:
+    void fromXmlNode(const XMLNode* node) final;
+
+    std::string get(const std::string& username, int online_id);
+};
+
+#endif // AUTH_RESOURCE_HPP
