+++ src/utils/command_manager/file_resource.hpp
@@ -1,0 +1,47 @@
+//
+//  SuperTuxKart - a fun racing game with go-kart
+//  Copyright (C) 2025 kimden
+//
+//  This program is free software; you can redistribute it and/or
+//  modify it under the terms of the GNU General Public License
+//  as published by the Free Software Foundation; either version 3
+//  of the License, or (at your option) any later version.
+//
+//  This program is distributed in the hope that it will be useful,
+//  but WITHOUT ANY WARRANTY; without even the implied warranty of
+//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//  GNU General Public License for more details.
+//
+//  You should have received a copy of the GNU General Public License
+//  along with this program; if not, write to the Free Software
+//  Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
+
+
+#ifndef FILE_RESOURCE_HPP
+#define FILE_RESOURCE_HPP
+
+#include "utils/types.hpp"
+#include "utils/command_manager/abstract_resource.hpp"
+#include <string>
+
+struct FileResource: public AbstractResource
+{
+private:
+    std::string      m_file_name;
+    uint64_t         m_interval;
+    mutable uint64_t m_last_invoked;
+    std::string      m_contents;
+
+    void tryUpdate();
+
+protected:
+    virtual void read();
+
+public:
+    void fromXmlNode(const XMLNode* node) final;
+
+    std::string get();
+};
+
+
+#endif // FILE_RESOURCE_HPP
